Proper working phase of all four service to place the order

+++ Ecom-Microservice/order-service/src/main/java/com/ecom/orderservice/model/OrderLineItems.java
@@ -9,6 +9,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @AllArgsConstructor
 @NoArgsConstructor
@@ -19,6 +21,6 @@ public class OrderLineItems {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long orderLineItemId;
     private String skuCode;
-    private Long price;
-    private int quantity;
+    private BigDecimal price;
+    private Integer quantity;
 }
